requirements.txt
langchain-pinecone==0.2.2
langchain-openai==0.3.0
langchain-community==0.3.14
python-dotenv==1.0.1
langchain-text-splitters==0.3.5
nemoguardrails==0.11.0
pypdf==5.1.0